renamed static/ to utils/
